# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Finance",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reconciliations/candidates": {
      "get": {
        "tags": [
          "reconciliation"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_api_v1_reconciliations_candidates_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Account Id"
            }
          },
          {
            "name": "cutoff_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Cutoff Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateResponse"
                  },
                  "title": "Response List Candidates Api V1 Reconciliations Candidates Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconciliations/preview": {
      "post": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Preview",
        "operationId": "preview_api_v1_reconciliations_preview_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconciliations": {
      "post": {
        "tags": [
          "reconciliation"
        ],
        "summary": "Complete",
        "operationId": "complete_api_v1_reconciliations_post",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconciliationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/economic": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Economic Report",
        "operationId": "economic_report_api_v1_reports_economic_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EconomicReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/cash-flow": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Cash Flow Report",
        "operationId": "cash_flow_report_api_v1_reports_cash_flow_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/net-worth": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Net Worth Report",
        "operationId": "net_worth_report_api_v1_reports_net_worth_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "as_of",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetWorthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateResponse": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "eligibility_date": {
            "type": "string",
            "format": "date",
            "title": "Eligibility Date"
          },
          "effect_cents": {
            "type": "integer",
            "title": "Effect Cents"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "transaction_id",
          "eligibility_date",
          "effect_cents"
        ],
        "title": "CandidateResponse"
      },
      "CashFlowReportResponse": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "receipts_cents": {
            "type": "integer",
            "title": "Receipts Cents"
          },
          "payments_cents": {
            "type": "integer",
            "title": "Payments Cents"
          },
          "net_cash_flow_cents": {
            "type": "integer",
            "title": "Net Cash Flow Cents"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionResponse"
            },
            "type": "array",
            "title": "Contributions"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "receipts_cents",
          "payments_cents",
          "net_cash_flow_cents",
          "contributions"
        ],
        "title": "CashFlowReportResponse"
      },
...
      "ContributionResponse": {
        "properties": {
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id"
          },
          "amount_cents": {
            "type": "integer",
            "title": "Amount Cents"
          },
          "economic_date": {
            "type": "string",
            "format": "date",
            "title": "Economic Date"
          },
          "cash_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cash Date"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "transaction_id",
          "amount_cents",
          "economic_date",
          "cash_date"
        ],
        "title": "ContributionResponse"
      },
...
      "EconomicReportResponse": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "income_cents": {
            "type": "integer",
            "title": "Income Cents"
          },
          "expense_cents": {
            "type": "integer",
            "title": "Expense Cents"
          },
          "result_cents": {
            "type": "integer",
            "title": "Result Cents"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionResponse"
            },
            "type": "array",
            "title": "Contributions"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "income_cents",
          "expense_cents",
          "result_cents",
          "contributions"
        ],
        "title": "EconomicReportResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NetWorthResponse": {
        "properties": {
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "assets_cents": {
            "type": "integer",
            "title": "Assets Cents"
          },
          "liabilities_cents": {
            "type": "integer",
            "title": "Liabilities Cents"
          },
          "net_worth_cents": {
            "type": "integer",
            "title": "Net Worth Cents"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/ContributionResponse"
            },
            "type": "array",
            "title": "Contributions"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "as_of",
          "assets_cents",
          "liabilities_cents",
          "net_worth_cents",
          "contributions"
        ],
        "title": "NetWorthResponse"
      },
...
      "ReconciliationRequest": {
        "properties": {
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "cutoff_date": {
            "type": "string",
            "format": "date",
            "title": "Cutoff Date"
          },
          "actual_balance_cents": {
            "type": "integer",
            "title": "Actual Balance Cents"
          },
          "selected_entry_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Entry Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "account_id",
          "cutoff_date",
          "actual_balance_cents",
          "selected_entry_ids"
        ],
        "title": "ReconciliationRequest"
      },
      "ReconciliationResponse": {
        "properties": {
          "reconciliation_id": {
            "type": "string",
            "title": "Reconciliation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "cutoff_date": {
            "type": "string",
            "format": "date",
            "title": "Cutoff Date"
          },
          "actual_balance_cents": {
            "type": "integer",
            "title": "Actual Balance Cents"
          },
          "prior_completed_cents": {
            "type": "integer",
            "title": "Prior Completed Cents"
          },
          "selected_effect_cents": {
            "type": "integer",
            "title": "Selected Effect Cents"
          },
          "checked_balance_cents": {
            "type": "integer",
            "title": "Checked Balance Cents"
          },
          "difference_cents": {
            "type": "integer",
            "title": "Difference Cents"
          },
          "selected_entry_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Selected Entry Ids"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          }
        },
        "type": "object",
        "required": [
          "reconciliation_id",
          "status",
          "account_id",
          "cutoff_date",
          "actual_balance_cents",
          "prior_completed_cents",
          "selected_effect_cents",
          "checked_balance_cents",
          "difference_cents",
          "selected_entry_ids"
        ],
        "title": "ReconciliationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-pf_session"
      }
    }
  }
}